x86: Secure Launch kernel early boot stub
The Secure Launch (SL) stub provides the entry point for Intel TXT to
jump to during the dynamic launch. The symbol sl_stub_entry is that
entry point and its offset into the kernel is conveyed to the
launching code using the Measured Launch Environment (MLE) header
in the structure named mle_header. The offset of the MLE header is set
in the kernel_info.

The startup SL routines (in sl_stub.S) contain the very early dynamic
launch setup code responsible for setting up the basic operating
environment to allow the normal kernel startup_32 code to proceed.
It is also responsible for properly waking and handling the APs on
Intel platforms.

The routine sl_main() runs after entering 64b mode in the setup
kernel. It is responsible for measuring configuration and module
information before it is used by the kernel. An example of entities
measured on Intel x86 are the boot params, the kernel command line,
the TXT heap, any external initramfs, etc. In addition this routine
does some early setup and validation of the environment like locating
the TPM event log and validating the location of various buffers to
ensure they are protected and not overlapping.

Signed-off-by: Daniel P. Smith <dpsmith@apertussolutions.com>
Signed-off-by: Ross Philipson <ross.philipson@oracle.com>