tpm: better locality handling
Commit 933bfc5ad213 introduced the use of a locality counter to control when
locality request was actually sent to the TPM. With this introduction, it
resulted in a hard enforcement of an assumption that the TPM had no active
localities at the time of the TPM driver initialization.  Couple the fact that
this may not always be the case and that the commit indiscriminately
decremented the counter created the condition for integer underflow of the
counter. When the underflow was triggered, it was early in tpm_tis_init_core
and all subsequent calls to request and relinquished calls would have the
counter flipping between underflow value and 0. The result is that it appeared
all calls to tpm_tis_request_locality() were successful, but they were not.
 The following call to tpm_tis_relinquish_locality() would decrement,
underflowing the counter, which would then not send the relinquish. The end
result is that the locality for which the TPM had active with the driver
loaded, would always remain active, to include after the driver shutdown.  This
creates a significant issue when using Intel TXT and Locality 2 was activated.
After the GETSEC[SEXIT] instruction is called, the PCH will close access to
Locality 2 MMIO address space, leaving the TPM locked in Locality 2 with no
means to relinquish the locality until system reset.

This commit seeks to address this situation through three primary changes. The
first is to walk the localities during initialization and close any open
localities. Next was to put a guard around the counter in the relinquish
function to ensure that it would not underflow. The last change was to make the
request locality functions to be consistent in their return values. The result
being that the functions will either return the locality requested if
successful or a negative error code.

Signed-off-by: Daniel P. Smith <dpsmith@apertussolutions.com>
Signed-off-by: Ross Philipson <ross.philipson@oracle.com>
Reported-by: Kanth Ghatraju <kanth.ghatraju@oracle.com>
Fixes: 933bfc5ad213 ("tpm, tpm: Implement usage counter for locality")