tpm: make locality request return value consistent
The return value and checks of the locality request functions are inconsistent.
A successful locality request will return the locality value that was
requested. This may be either a zero or a positive number. Checks of boolean
not (!) of the return value will not only catch negative error codes, but also
localities greater than zero.

This commit reworks the locality request functions to return either the
locality or an appropriate negative error code. Included in addressing the
correct error code, an added check to enforce only a valid locality. At the
call sites of the locality requests, return value checks were updated to
checking for negative returns instead of a boolean not.

Signed-off-by: Daniel P. Smith <dpsmith@apertussolutions.com>
Signed-off-by: Ross Philipson <ross.philipson@oracle.com>